Use 64 Bit Int For Thread Id

+++ c_impl/src/util/script_thread.h
@@ -5,7 +5,7 @@
 
 /// A function pointer for when the script instance has closed and the thread is no longer needed.
 typedef void(*CubsThreadOnScriptClose)(void* threadObj);
-typedef uint32_t(*CubsThreadGetId)(const void* threadObj);
+typedef uint64_t(*CubsThreadGetId)(const void* threadObj);
 typedef void(*CubsThreadClose)(void* threadObj);
 
 typedef struct CubsThreadVTable {
@@ -26,4 +26,4 @@ CubsThread cubs_thread_spawn(bool closeWithScript);
 
 void cubs_thread_close(CubsThread* thread);
 
-uint32_t cubs_thread_get_id(const CubsThread* thread);
+uint64_t cubs_thread_get_id(const CubsThread* thread);
